refactor(constants): size와 category items를 하나의 상수 파일로 통합
