+++ constants/index.ts
@@ -1,4 +1,3 @@
 // 상수명 중복을 방지하고 모듈 관리를 위해 index.ts에서 re-export 합니다.
-export * from './category-items';
+export * from './items';
 export * from './route-paths';
-export * from './size-items';
