+++ constants/items.ts
@@ -1,4 +1,5 @@
-// TODO: 백엔드 데이터와 value 값 일치 시키기
+// 서버와 프론트엔드 모두에서 사용하는 value-label 매핑 상수입니다.
+
 export const CATEGORY_ITEMS = [
   { label: '섬유 & 텍스타일 아트', value: 'TEXTILE_ART' },
   { label: '시각 예술', value: 'VISUAL_ART' },
@@ -8,4 +9,11 @@ export const CATEGORY_ITEMS = [
   { label: '가구 & 인테리어', value: 'FURNITURE_INTERIOR' },
   { label: '자연 & 생활 공예', value: 'NATURE_LIFE_CRAFT' },
   { label: '패션 & 의류', value: 'FASHION_CLOTHING' },
+] as const;
+
+export const SIZE_ITEMS = [
+  { label: '대형', value: 'L' },
+  { label: '중형', value: 'M' },
+  { label: '소형', value: 'S' },
+  { label: '해당없음', value: 'X' },
 ] as const;
