--- constants/size-items.ts
@@ -1,7 +1,0 @@
-// TODO: 백엔드 데이터와 value 값 일치 시키기
-export const SIZE_ITEMS = [
-  { label: '대형', value: 'L' },
-  { label: '중형', value: 'M' },
-  { label: '소형', value: 'S' },
-  { label: '해당없음', value: 'X' },
-] as const;
